Introduced the concept of GameAsset and a helper method for decompilers
The decompilers will usually be iterating over a game folder, decompiling all assets inside. So this helper takes that folder and returns the enumeration of game assets (already deserialized). Decompilers merely need to transform from the raw deserialization into a proper type and then save it accordingly.

Note that in the future custom XNA serializers/deserializers may simply be used with some automatic decompilation. The benefit here is that those serializers/deserializers could be used in other projects, perhaps even in the game itself (which would probably clean it up a bit?)